Delegate tableview’s delegate implementation to the view model
