+++ WooCommerce/Classes/ViewModels/OrderDetailsViewModel/OrderDetailsViewModel.swift
@@ -2,8 +2,9 @@ import Foundation
 import UIKit
 import Gridicons
 import Yosemite
+import MessageUI
 
-class OrderDetailsViewModel {
+final class OrderDetailsViewModel {
     let order: Order
     let orderStatus: OrderStatus?
     let currencyFormatter = CurrencyFormatter()
@@ -226,7 +227,7 @@ class OrderDetailsViewModel {
 
     /// Sections to be rendered
     ///
-    var sections = [Section]()
+    private var sections = [Section]()
 
     /// Order Notes
     ///
@@ -247,6 +248,10 @@ class OrderDetailsViewModel {
 //    var dataSource: UITableViewDataSource {
 //        return self
 //    }
+
+    private let emailComposer = OrderEmailComposer()
+
+    private let messageComposer = OrderMessgeComposer()
 }
 
 
@@ -908,6 +913,116 @@ extension OrderDetailsViewModel {
 }
 
 
+extension OrderDetailsViewModel {
+    func tableView(_ tableView: UITableView,
+                   in viewController: UIViewController,
+                   didSelectRowAt indexPath: IndexPath) {
+        switch sections[indexPath.section].rows[indexPath.row] {
+
+        case .addOrderNote:
+            WooAnalytics.shared.track(.orderDetailAddNoteButtonTapped)
+
+            let newNoteViewController = NewNoteViewController()
+            newNoteViewController.viewModel = self
+
+            let navController = WooNavigationController(rootViewController: newNoteViewController)
+            viewController.present(navController, animated: true, completion: nil)
+        case .trackingAdd:
+            WooAnalytics.shared.track(.orderDetailAddTrackingButtonTapped)
+
+            let addTrackingViewModel = AddTrackingViewModel(order: order)
+            let addTracking = ManualTrackingViewController(viewModel: addTrackingViewModel)
+            let navController = WooNavigationController(rootViewController: addTracking)
+            viewController.present(navController, animated: true, completion: nil)
+        case .orderItem:
+            let item = order.items[indexPath.row]
+            let productID = item.variationID == 0 ? item.productID : item.variationID
+            let loaderViewController = ProductLoaderViewController(productID: productID,
+                                                                   siteID: order.siteID)
+            let navController = WooNavigationController(rootViewController: loaderViewController)
+            viewController.present(navController, animated: true, completion: nil)
+        case .details:
+            WooAnalytics.shared.track(.orderDetailProductDetailTapped)
+            viewController.performSegue(withIdentifier: Constants.productDetailsSegue, sender: nil)
+        case .billingEmail:
+            WooAnalytics.shared.track(.orderDetailCustomerEmailTapped)
+            displayEmailComposerIfPossible(from: viewController)
+        case .billingPhone:
+            displayContactCustomerAlert(from: viewController)
+        default:
+            break
+        }
+    }
+}
+
+
+// MARK: - MFMailComposeViewControllerDelegate Conformance
+//
+private extension OrderDetailsViewModel {
+    func displayEmailComposerIfPossible(from: UIViewController) {
+        emailComposer.displayEmailComposerIfPossible(for: order, from: from)
+    }
+
+    /// Displays an alert that offers several contact methods to reach the customer: [Phone / Message]
+    ///
+    func displayContactCustomerAlert(from: UIViewController) {
+        guard let sourceView = from.view else {
+            return
+        }
+
+        let actionSheet = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
+        actionSheet.view.tintColor = StyleManager.wooCommerceBrandColor
+
+        actionSheet.addCancelActionWithTitle(ContactAction.dismiss)
+        actionSheet.addDefaultActionWithTitle(ContactAction.call) { [weak self] _ in
+            guard let self = self else {
+                return
+            }
+
+            guard let phoneURL = self.order.billingAddress?.cleanedPhoneNumberAsActionableURL else {
+                return
+            }
+
+            WooAnalytics.shared.track(.orderDetailCustomerPhoneOptionTapped)
+            self.callCustomerIfPossible(at: phoneURL)
+        }
+
+        actionSheet.addDefaultActionWithTitle(ContactAction.message) { [weak self] _ in
+            WooAnalytics.shared.track(.orderDetailCustomerSMSOptionTapped)
+            self?.displayMessageComposerIfPossible(from: from)
+        }
+
+        let popoverController = actionSheet.popoverPresentationController
+        popoverController?.sourceView = sourceView
+        popoverController?.sourceRect = sourceView.bounds
+
+        from.present(actionSheet, animated: true)
+
+        WooAnalytics.shared.track(.orderDetailCustomerPhoneMenuTapped)
+    }
+
+    /// Attempts to perform a phone call at the specified URL
+    ///
+    func callCustomerIfPossible(at phoneURL: URL) {
+        guard UIApplication.shared.canOpenURL(phoneURL) else {
+            return
+        }
+
+        UIApplication.shared.open(phoneURL, options: [:], completionHandler: nil)
+        WooAnalytics.shared.track(.orderContactAction, withProperties: ["id": order.orderID,
+                                                                        "status": self.order.statusKey,
+                                                                        "type": "call"])
+
+    }
+}
+
+private extension OrderDetailsViewModel {
+    func displayMessageComposerIfPossible(from: UIViewController) {
+        messageComposer.displayMessageComposerIfPossible(order: order, from: from)
+    }
+}
+
+
 // MARK: - Syning data. Yosemite related stuff
 extension OrderDetailsViewModel {
     func syncOrder(onCompletion: ((Order?, Error?) -> ())? = nil) {
@@ -1008,3 +1123,18 @@ extension OrderDetailsViewModel {
         StoresManager.shared.dispatch(deleteTrackingAction)
     }
 }
+
+private extension OrderDetailsViewModel {
+    enum ContactAction {
+        static let dismiss = NSLocalizedString("Dismiss", comment: "Dismiss the action sheet")
+        static let call = NSLocalizedString("Call", comment: "Call phone number button title")
+        static let message = NSLocalizedString("Message", comment: "Message phone number button title")
+    }
+
+    enum Constants {
+        static let productDetailsSegue = "ShowProductListViewController"
+        static let orderStatusListSegue = "ShowOrderStatusListViewController"
+    }
+}
+
+
